initial commit
This plugin houses a single function, serving to provide dwim
autocompletion. It's most of the way there -- the only glaring
defficiency I'm currently aware of is the treatment of buffers where
noexpandtab is set, in which case it always uses the Tab key to insert
tab characters, never for completion.